Add `McStateExtra`

+++ proc/src/derive_load.rs
@@ -121,7 +121,7 @@ fn build_struct(
 }
 
 fn load_tag_op(tag: attr::TlbTag) -> Option<TokenStream> {
-    let bits = tag.bits;
+    let bits = tag.bits as u16;
 
     let neg_condition = match bits {
         0 => return None,
